+++ Nahyeon/src/test/java/com/depromeet/nahyeon/common/controller/HealthCheckControllerTest.java
@@ -1,4 +1,4 @@
-package com.depromeet.nahyeon.controller;
+package com.depromeet.nahyeon.common.controller;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
