+++ Nahyeon/src/test/java/com/depromeet/nahyeon/post/controller/PostControllerTest.java
@@ -1,4 +1,4 @@
-package com.depromeet.nahyeon.controller;
+package com.depromeet.nahyeon.post.controller;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
