+++ Nahyeon/src/test/java/com/depromeet/nahyeon/user/controller/UserCreateControllerTest.java
@@ -1,20 +1,25 @@
-package com.depromeet.nahyeon.controller;
+package com.depromeet.nahyeon.user.controller;
 
+import static org.mockito.ArgumentMatchers.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import org.junit.jupiter.api.Test;
+import org.mockito.BDDMockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.SqlGroup;
 import org.springframework.test.web.servlet.MockMvc;
 
-import com.depromeet.nahyeon.post.domain.PostCreate;
+import com.depromeet.nahyeon.user.domain.UserCreate;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @SpringBootTest
@@ -22,10 +27,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 @AutoConfigureTestDatabase
 @TestPropertySource("classpath:test-application.yml")
 @SqlGroup({
-	@Sql(value = "/sql/post-create-controller-test-data.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD),
 	@Sql(value = "/sql/delete-all-data.sql", executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
 })
-class PostCreateControllerTest {
+public class UserCreateControllerTest {
 
 	@Autowired
 	private MockMvc mockMvc;
@@ -33,23 +37,28 @@ class PostCreateControllerTest {
 	@Autowired
 	private ObjectMapper objectMapper;
 
+	@MockBean
+	private JavaMailSender mailSender;
+
 	@Test
-	void 사용자는_게시물을_작성할_수_있다() throws Exception {
+	void 사용자는_회원_가입을_할_수_있고_회원가입된_사용자는_PENDING_상태이다() throws Exception {
 		// given
-		PostCreate postCreateDto = PostCreate.builder()
-			.content("post content")
-			.writerId(1)
+		UserCreate userCreateDto = UserCreate.builder()
+			.email("nahyeon@kakao.com")
+			.nickname("hyeon")
+			.address("Pangyo")
 			.build();
+		BDDMockito.doNothing().when(mailSender).send(any(SimpleMailMessage.class));
 
 		// when
 		// then
-		mockMvc.perform(post("/api/posts")
+		mockMvc.perform(post("/api/users")
 				.contentType(MediaType.APPLICATION_JSON)
-				.content(objectMapper.writeValueAsString(postCreateDto)))
+				.content(objectMapper.writeValueAsString(userCreateDto)))
 			.andExpect(status().isCreated())
-			.andExpect(jsonPath("$.id").value(1))
-			.andExpect(jsonPath("$.content").value("post content"))
-			.andExpect(jsonPath("$.writer.id").isNumber())
-			.andExpect(jsonPath("$.writer.nickname").isString());
+			.andExpect(jsonPath("$.id").isNumber())
+			.andExpect(jsonPath("$.email").value("nahyeon@kakao.com"))
+			.andExpect(jsonPath("$.nickname").value("hyeon"))
+			.andExpect(jsonPath("$.status").value("PENDING"));
 	}
 }
